// frontend/web/src/services/API/proserve-wb-provisioning-api/models/ComponentVersionDetail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ComponentVersionDetail
 */
export interface ComponentVersionDetail {
    /**
     * The component version name.
     * @type {string}
     * @memberof ComponentVersionDetail
     */
    componentName: string;
    /**
     * The component version type (MAIN|HELPER).
     * @type {string}
     * @memberof ComponentVersionDetail
     */
    componentVersionType: string;
    /**
     * The software vendor name.
     * @type {string}
     * @memberof ComponentVersionDetail
     */
    softwareVendor: string;
    /**
     * 
     * @type {string}
     * @memberof ComponentVersionDetail
     */
    softwareVersion: string;
    /**
     * The license dashboard for the software vendors.
     * @type {string}
     * @memberof ComponentVersionDetail
     */
    licenseDashboard?: string;
    /**
     * Notes about the component version.
     * @type {string}
     * @memberof ComponentVersionDetail
     */
    notes?: string;
}

/**
 * Check if a given object implements the ComponentVersionDetail interface.
 */
export function instanceOfComponentVersionDetail(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentName" in value;
    isInstance = isInstance && "componentVersionType" in value;
    isInstance = isInstance && "softwareVendor" in value;
    isInstance = isInstance && "softwareVersion" in value;

    return isInstance;
}

export function ComponentVersionDetailFromJSON(json: any): ComponentVersionDetail {
    return ComponentVersionDetailFromJSONTyped(json, false);
}

export function ComponentVersionDetailFromJSONTyped(json: any, ignoreDiscriminator: boolean): ComponentVersionDetail {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentName': json['componentName'],
        'componentVersionType': json['componentVersionType'],
        'softwareVendor': json['softwareVendor'],
        'softwareVersion': json['softwareVersion'],
        'licenseDashboard': !exists(json, 'licenseDashboard') ? undefined : json['licenseDashboard'],
        'notes': !exists(json, 'notes') ? undefined : json['notes'],
    };
}

export function ComponentVersionDetailToJSON(value?: ComponentVersionDetail | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'componentName': value.componentName,
        'componentVersionType': value.componentVersionType,
        'softwareVendor': value.softwareVendor,
        'softwareVersion': value.softwareVersion,
        'licenseDashboard': value.licenseDashboard,
        'notes': value.notes,
    };
}